requirements.txt
fastapi[all]
fastapi-cache2
redis